requirements.txt
# ──────────────────────────────────────────────────────────────────────────────
# requirements.txt — ICO Distribuidora v2 (NeuMF)
# Python 3.10+
# ──────────────────────────────────────────────────────────────────────────────

# ── ML — NeuMF ────────────────────────────────────────────────────────────────
torch>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0     # roc_auc_score en evaluate.py
pyarrow>=14.0.0         # lectura/escritura parquet

# ── API backend ───────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0   # JWT
bcrypt>=4.0.0                      # hash de contraseñas
python-multipart>=0.0.9            # form data (login)

# ── Base de datos ─────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0
aiosqlite>=0.19.0       # SQLite async para FastAPI

# ── Notebooks ─────────────────────────────────────────────────────────────────
jupyter>=1.0.0
ipykernel>=6.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# ── Serialización ─────────────────────────────────────────────────────────────
joblib>=1.3.0

# ── Utilidades ────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0    # variables de entorno
requests>=2.31.0        # llamadas HTTP entre servicios
